add fetch timeout tests
