+++ test/scene-system/sdk/index.ts
@@ -1,1 +1,2 @@
-import './usingFetchAndWebsocket.spec'
+import './websocketWrapper.spec'
+import './fetchWrapper.spec'
